Made no MPI version of oceansar_sarsim the default one. The idea is to deprecate oceansar_sarsim_mpi. Updating some plotting rutines in sar_processor (mainly for debugging purposes)
